% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{schemaAppend}
\alias{schemaRemove}
\title{Schema Modifiers}
\usage{
schemaAppend(schema, append)

schemaRemove(schema, remove)
}
\arguments{
\item{schema}{\code{structType}. A DataFrame's schema, see \code{\link[SparkR:schema]{SparkR::schema()}}.}

\item{append}{\code{list()}. A \code{list} of \code{\link[SparkR:structField]{SparkR::structField()}}s to append to the
end of the DataFrame.}

\item{remove}{\code{character(n)}. A vector of field names to be removed from the
\code{schema}.}
}
\description{
\emph{schemaAppend}: Given a DataFrame's schema, append additional
fields to it.

\emph{schemaRemove}: Given a DataFrame's schema, remove fields from
it.
}
\examples{
\dontrun{
x <- SparkR::sql("SELECT * FROM tbl")
newCol <- SparkR::structField("newCol", "double")
newSchema <- schemaAppend(
  schema = SparkR::schema(x),
  fields = list(newCol)
)
newSchema

schemaRemove(
  schema = newSchema,
  remove = "newCol"
)
}

}
\seealso{
\code{\link[SparkR:structType]{SparkR::structType()}}, \code{\link[SparkR:structField]{SparkR::structField()}}, \code{\link[SparkR:schema]{SparkR::schema()}}
}
