% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaAppend.R
\name{schemaAppend}
\alias{schemaAppend}
\title{Append Additional Fields to a Schema}
\usage{
schemaAppend(schema, fields)
}
\arguments{
\item{schema}{\code{structType}. A DataFrame's schema, see \code{\link[SparkR:schema]{SparkR::schema()}}.}

\item{fields}{\code{list()}. A \code{list} of \code{\link[SparkR:structField]{SparkR::structField()}}s to append to the
end of the DataFrame.}
}
\description{
Given a DataFrame's schema, append additional fields to it.
}
\examples{
\dontrun{
x <- SparkR::sql("SELECT * FROM tbl")
new_col <- SparkR::structField("new_col", "double")
schemaAppend(
  schema = SparkR::schema(x),
  fields = list(new_col)
)
})

}
\seealso{
\code{\link[SparkR:structType]{SparkR::structType()}}, \code{\link[SparkR:structField]{SparkR::structField()}}, \code{\link[SparkR:schema]{SparkR::schema()}}
}
